# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for telemetry service",
        "title": "Telemetry API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/telemetry/v1/data/latest": {
            "get": {
                "description": "Retrieve the latest properties/attributes for a specific device in a space. Organization is resolved from X-Organization header or hostname (e.g., {org}.localhost)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "data"
                ],
                "summary": "Get device properties",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device ID",
                        "name": "device_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Space slug",
                        "name": "space_slug",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Device properties as key-value pairs",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/telemetry/v1/entities": {
            "get": {
                "description": "Retrieve a paginated list of entities with optional filtering. Organization is resolved from X-Organization header or hostname (e.g., {org}.localhost)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "entities"
                ],
                "summary": "Get entities",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by entity category",
                        "name": "category",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by device ID",
                        "name": "device_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by display type (comma-separated)",
                        "name": "display_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search term for filtering",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of results per page (default 20)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of results to skip (default 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.PaginatedResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/telemetry/v1/events/device/{device_id}": {
            "get": {
                "description": "Retrieve all events for a specific device with optional time range filtering. Organization is resolved from X-Organization header or hostname (e.g., {org}.localhost)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Get events by device",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device ID",
                        "name": "device_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Maximum number of events to return (default 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Start time as Unix timestamp (milliseconds)",
                        "name": "start_time",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "End time as Unix timestamp (milliseconds)",
                        "name": "end_time",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.EventsByDeviceResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "common.PaginatedResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "next": {
                    "type": "string"
                },
                "previous": {
                    "type": "string"
                },
                "results": {}
            }
        },
...
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "models.EventItem": {
            "type": "object",
            "properties": {
                "entity_id": {
                    "type": "string"
                },
                "event_data": {
                    "type": "object",
                    "additionalProperties": true
                },
                "event_id": {
                    "type": "integer"
                },
                "event_level": {
                    "type": "string"
                },
                "event_type": {
                    "type": "string"
                },
                "space_slug": {
                    "type": "string"
                },
                "time_fired": {
                    "type": "string"
                }
            }
        },
...
        "models.EventsByDeviceResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "device_id": {
                    "type": "string"
                },
                "events": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.EventItem"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "API Key": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "Access Token": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "Organization slug name": {
            "type": "apiKey",
            "name": "X-Organization",
            "in": "header"
        },
        "Space slug name": {
            "type": "apiKey",
            "name": "X-Space",
            "in": "header"
        }
    }
}